Use TLS serialization
The serialization relies on the `tls_codec_derive` derive macros and a set of adapter types that can be composed to support serializing foreign types not supported by `tls_codec` out of the box. These adapters are defined in the `tls` module.

`usize`, being platform-dependent and not supported by `tls_codec`, was replaced by `u32`.

Note that this commit has gaps with regards to compliance with the MLS RFC, especially regarding enumeration discriminant sizes. A subequent task shall go through them all and ensure that the right `repr` attribute is used to comply with the RFC.

Fixes #8