Improve TLS serialization performance
Based on criterion results, this PR improves the performance of both serializing and deserializing sequences of `SecretKey`s. Similar gains should apply to any type whose `tls_serialized_len` implementation is inefficient (e.g. it needs to serialize to figure out the length).

Note that `Vec::with_capacity` was changed to `Vec::new` as the available length was in bytes, not number of items.