Remove unused dependencies
This removes the unused enum-iterator, base64, num_enum, indexmap, and
sha1 dependencies. The enum-iterator was mentioned in the code (but
not used), the others were not mentioned at all outside of the
Cargo.toml files.

Using

    cargo tree -p mls-rs -e normal --prefix none | sort -u | wc -l

to count unique transitive dependencies of mls-rs, this reduces the
number from 60 to just 45.