Customize debug output of byte arrays
This should help reduce the noise when troubleshooting issues and
debug-printing values.

By default byte arrays are written with their length only, unless the
alternate debug format is requested (`"{:#?}"`), in which case the bytes
are also written in hex format. Group IDs should usually be short and
their bytes are always written.