Execute UniFFI bindings via a simple Python test (#87)
This uses the UniFFI test helpers to make the `libmls_rs_uniffi.so`
file available to the Python script.

I would have liked to use `$CARGO_TARGET_TMPDIR` to place the
temporary files in a stable place under `target/`, but this variable
is only set for integration tests, not unit tests. So the script now
ends up somewhere in your system temp directory.

We could decide to only use integration tests for these tests, let me
know what you think!

The test is very simple here: it simply executes the Python script,
which then has to return with a zero exit code. If it fails, the
output is printed and you then have to debug the Python code by hand.

Related to #81.