Add a `generate_python_tests!` macro
The macro generates tests for Python scripts found next to the
integration test.

This saves a bit of boiler-plate code at the point where the test is
created. It also allows us to quickly see if there are gaps in the
test coverage between the sync and async code.

This comes at the expense of more “magic” in the form of the macro.