Add a `generate_python_tests!` macro
The macro takes the names of one or two Python scripts inside the
`test_bindings/` directory and uses them to construct unit tests.

This saves a bit of boiler-plate code at the point where the test is
created. It also allows us to quickly see if there are gaps in the
test coverage between the sync and async code.

This comes at the expense of more “magic” in the form of the macro.
The consistency of no boiler-plate code might make this worth the
cost.