Add `client_config_default` function for easy client configuration (#111)
* [uniffi] Merge `FFICallbackError` into `Error`

Part of the goal of mls-rs-uniffi is to present a small FFI API.
Currently, it does not seem to be necessary to use a separate error
type here.

* [uniffi] Move `GroupStateStorage` wrapper to client

This type is specific to our client type so it can be nicely grouped
with it in the client module.

* [uniffi] Add a `client_config_default` function

This function returns a client config with some simple defaults. Right
now, the default is in-memory storage.

The new `GroupStateStorageAdapter` struct allows us to use any mls-rs
group state storage, so we could easily surface the Sqlite storage as
well now.