[uniffi] Remove `KeyPackage` from the API
We don’t think people will need to access the key packages and their
fields. As an example, when adding members to a group, you use
messages, not key packages.

Related to #81, reverts most of #102. I left the `ProtocolVersion`
behind, but I can remove it too if people think we don’t need it.