[uniffi] Remove `GroupState` type from storage trait
This removes the `GroupState` type from the `GroupStateStorage` trait
in mls-rs-uniffi. This is done for two reasons:

- It makes all methods on the trait consistent in the way that they
  take a `group_id` as the first argument.

- Fewer types means less boiler-plate generated by the UniFFI bindings
  generator (`Vec<u8>` is a built-in type in UniFFI).

The `EpochRecord` type stays: I don’t think we have a different way of
passing in a vector of epochs IDs with their data.

I could apply this change to mls-rs too, but it doesn’t seem useful
there since it’s cheap and easy to create a new type in Rust.