[uniffi] Setup CI for the Kotlin code
This sets up a simple CI pipeline for the generated Kotlin code on
Linux and macOS. This would have allowed us to find and fix #101 and
mozilla/uniffi-rs#2032 sooner.

This updates Git commit of the UniFFI dependency to the merge of
mozilla/uniffi-rs#2032.

I found that I need to use the beta macOS 13 GitHub runner instead of
the “latest” macOS 12 runner used by default.

The download of the Kotlin dependencies doesn’t work on Windows, so
the integration tests only run on Linux and macOS right now.

Fixes #103.