Add more caching to the CI
This is a followup to #137 to make the caching more consistent across
the different CI jobs.

I’ve noticed from other projects that it’s not very useful to cache
the build artifacts for each branch: that just ends up polluting the
10 GB cache GitHub provides us with. So saving the cache is now
disabled on non-main branches (the cache will still be read on all PRs
from all branches).