[uniffi] Introduce a `GroupId` newtype wrapper
This case up during our API review: instead of exposing a `Vec<u8>`,
we can let the API use a `GroupId` newtype wrapper.

A downside of this is that UniFFI doesn’t support exposing traits like
`Hash` on its record types (see
https://github.com/mozilla/uniffi-rs/issues/2043). A consequence of
this is that the Python `GroupId` type is not hashable. I therefore
ended up accessing the `.id` attribute directly in the integration
test.

Related to #81.