[uniffi] Make mls-rs-crypto-openssl optional
For Android, we will be using a crypto provider based on BoringSSL. It
is therefore important that mls-rs-uniffi becomes agnostic to the
crypto provider.

This PR handles this very simply: by requiring two things:

- the user sets a Cargo feature to depend on the right crypto provider
  implementation,
- that the crypto provider has a suitable `new()` method that can be
  used to instantiate it.

A better alternative would be to make the choice of crypto provider
dynamic via a trait. Unfortunately, I don’t think this is supported
with UniFFI because it doesn’t support associated types in exported
traits and because it also doesn’t support generic exported types.

Making a single non-generic type which uses

```rust
struct CryptoProviderAdapter {
    crypto_provider: Box<dyn CryptoProvider>
}
```

fail because the `CryptoProvider` trait has an associated type, which
I cannot specify at this point.