[crypto-rustcrypto]: remove unused `custom` feature
The `custom` feature from `getrandom` is not actually used in normal
builds — infact, `getrandom` itself is never used directly.

However, the feature can be used by a binary on an embedded platform
to register a custom random number generator. The final binary should
enable the `custom` feature to do this, see the
`register_custem_genrandom!` macro for more details:

  https://docs.rs/getrandom/latest/getrandom/macro.register_custom_getrandom.html

We emulate this in the CI build by enabling the feature on the command
line.