Add support for last resort key packages
This introduces a `LastResortKeyPackageExt` extension which
corresponds to the last resort key package extension described here:

  https://datatracker.ietf.org/doc/html/draft-ietf-mls-extensions-04#name-last-resort-keypackages

The new code is hidden behind a `last_resort_key_package_ext` Cargo
feature.

There is not yet a super convenient way to set the extension on a
single key package: the key package extensions are fixed when the
client is created, so one has to create a temporary throw-away client
for this. This can be addressed in a future PR, if needed.