Format with `imports_granularity = "Module"`
This is the result of running `cargo +nightly fmt` with the unstable
`rustfmt` option `imports_granularity = "Module"`.

Personally, I find this style much easier to use:

- The lack of deep nesting makes it easier for me to understand the
  imports at a glance.
- Because most imports are on a single line, I can add/remove them
  easier. Copy-pasting between files is easier.

Of course, this option is unstable, meaning that it has no effect on a
stable `rustfmt`. Luckily, the default value for `imports_granularity`
is `Preserve`, which means that the formatting done here will be
preserved by future `cargo fmt` calls:

  https://github.com/rust-lang/rustfmt/blob/master/Configurations.md#imports_granularity

This applies to both on the stable and nightly channels.