Remove overloaded public methods from X509IdentityProvider (#236)
* Avoid X509IdentityProvider overloaded methods

While using a `X509IdentityProvider` in a generic context, I
discovered that it has a few methods which I found confusing:

- X509IdentityProvider::identity takes 1 argument, but
  IdentityProvider::identity takes 2 arguments!

- X509IdentityProvider::valid_successor takes 2 arguments,
  but IdentityProvider::valid_successor takes 3 arguments!

I ended up solving this with the fully-qualified `<Type as
Trait>::method` syntax, but I think the situation can be avoided all
together by just implementing the trait directly.

* Bump mls-rs-identity-x509 version to 0.14

The previous commit removed a few public methods, implying that we
need a SemVer breaking change. Since the types are re-exported in
mls-rs, I also bumped the version of that crate.

* Use imported names more consistently

---------

Co-authored-by: mulmarta <103590845+mulmarta@users.noreply.github.com>