Add support for P384 public key algorithm (#251)
* Add support for P384 public key algorithm

I found that I could not use root certificates which use a P384 public
key algorithm. More specifically, passing in a certificate with

```
$ openssl x509 -noout -text -in root.der
…
        Subject Public Key Info:
            Public Key Algorithm: id-ecPublicKey
                Public-Key: (384 bit)
                pub:
                    …
                ASN1 OID: secp384r1
                NIST CURVE: P-384
…
```

gives me

```
EcX509Error(UnsupportedPublicKeyAlgorithm("ObjectIdentifier(1.2.840.10045.2.1)"))
```

back. The changes here seem to fix this, but I’ll admit that they were
made very mechanically based on the existing code.

This would be a continuation of the work in #190.

* Add P384 certificate test

The test certificate was generated using certificatetools.com, which I
believe is where the others were generated as well.