Export key generation for encrypt_application_message and process_incoming_message.
Used by clients to authenticate the generation to defend against in-group forgery
attacks described in https://eprint.iacr.org/2025/554. This may be accomplished
by placing the generation in the `message` or `authenticated_data` parameters of
`Group::encrypt_application_message`, as both fields are signed by the sender's
signature key. To verify, the authenticated generation MUST be checked against
the unauthenticated generation in the SenderData, which is the actual value used
to derive keys for decryption.